Entitlement enforcement.
Hooked up entitlement to state machine.
Added halt and warn states to Entitlement, the expansion of which is dealt with by the EntitlementExpander. This made it easier to deal with at the state machine level.
Added entitlement middleware that deals with enforcement.
Fixed conversion of warn_at and halt_at to SystemTime (this was epoch secs rather than millis).
Added new Opt options.