Telemetry hot reload (#2358)
This PR is POC that reexamines the lifecycle of opentelemetry within the
router.

Closes #1812
Closes #2359 
Closes #2338 
Closes #2113

# Background
We previously had hot reload capability, however there were issues with
the implementation:
* The native tracing `ReloadLayer` was used, but this could not support
the `tracing-opentelemetry` layer due to the need to support layer
downcast. The reload layer is not compatible with any layer that
requires this.
* It was unclear what the requirements for disposing of tracing
exporters was and we did this during `Drop`.

# Current state
* Hot reload is disabled for telemetry. There is code in configuration
that checks if the telemetry config changed and prevents progressing of
state.
* Initialization of tracing happens in two places (set as dispatcher).
Once at application startup, and once in the telemetry plugin (set as
global).

# Issues
* Not having hot reload for telemetry is a bad user experience. We are
aware that users are using hot reload in production, and they will have
to recycle routers if they hit this.
* Shutdown of telemetry is happening prematurely on exit. This manifests
itself as `OpenTelemetry trace error occurred: cannot send span to the
batch span processor because the channel is closed`. This has been
reported by users, cloud and developers. (This happens every time I run
tests and causes the test suite to fail).
* There is singleton pattern initialization within the telemetry plugin.
Preventing reload of the telemetry plugin doesn't actually mean that the
plugin didn't get replaced, so it ends up in an inconsistent state that
is difficult to reason about.
* Metrics fail to work after hot reload. This is due to the metrics
exporters being tied to the telemetry plugin lifecycle, but the
singleton state not getting updated. It's not as simple as moving more
state into a singleton as we interact with other parts of the router
e.g. endpoints.
* There is code that has leaked into the configuration logic that has to
special case telemetry.

# What we are doing differently this time
* Bring lifecycle in line with recommendations for tracing and otel.
This means:
1. Tracing should be initialized once at application startup, and the
subscriber should not be replaced.
2. `tracing-subscriber::fmt` should use the tracing `ReloadLayer`. It is
compatible and but needs perf testing.
3. opentelemetry should be shut down at the end of execution. It is the
last thing that should happen and should not be tied to Drop. This is in
line with docs from otel.
4. `ReloadLayer` and derivatives will not be used for opentelemetry
layer.
* `TracerProvider`s will not be disposed of in `Drop` unless it is unused. This was the root
of many of our hanging problems as it was difficult to reason about when
this was being called. The Otel guidance for dropping `TracerProvider`s
is now clear, you do this from within a blocking task.
* Using a `ReloadTracer` rather than `ReloadLayer`. This is much simpler
than a reload layer as it doesn't require anything special around
downcasting. This can be passed to a regular
`tracing_opentelemetry::layer`.

## Benchmarks:
### New
```
basic_composition_benchmark
                        time:   [516.70 µs 526.25 µs 538.38 µs]
                        change: [-22.383% +6.3578% +44.889%] (p = 0.72 > 0.05)
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  5 (5.00%) high mild
  7 (7.00%) high severe
```
### Old
```
basic_composition_benchmark
                        time:   [467.11 µs 476.12 µs 487.72 µs]
                        change: [-26.421% +1.5537% +38.940%] (p = 0.93 > 0.05)
                        No change in performance detected.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe
```