chore: Phase 3 of changelog project
Probably not the final phase of this project, but another step in the right
direction of automating the changelog juggling.

This follows up on #2545 [1] and #2534 [2], which introduced an `xtask`
command for automatically creating a changeset file in the `.changesets`
directory.

With this addition, the existing release step (also `cargo xtask`) will now
roll-up those changesets and plop them into the `CHANGELOG.md`, obeying a
pre-prescribed order for how they should be presented and utilizing the file
prefixes to bucket them accordingly.

The details of the changesets behaviour is outlined in the README of the
`.changesets` directory, referenced below [3].

[1]: https://github.com/apollographql/router/pull/2545
[2]: https://github.com/apollographql/router/pull/2534
[3]: https://github.com/apollographql/router/blob/dev/.changesets/README.md