Metrics attributes should allow other types than string (#2616)
Convert Metrics attributes to `AttributeValue`. `AttributeValue`
encapsulates the full range of what metrics attributes can be.

Metrics attributes in OpenTelemetry allow the following types:
* string
* string[]
* float
* float[]
* int
* int[]
* bool
* bool[]

However, our configuration only allowed strings. This has been fixed,
and therefore it is now possible to use booleans via env expansion as
metrics attributes.

For example:
```yaml
telemetry:
  metrics:
    prometheus:
      enabled: true
    common:
      attributes:
        supergraph:
          static:
            - name: "my_boolean"
              value: '${env.MY_BOOLEAN:-true}'
```


Fixes #2510

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: bryn <bryn@apollographql.com>