remove requirement for Clone from AsyncCheckpoint
I was thinking about things a bit overnight in particular, about the
fact that Geoffroy had re-engineered the router plugin pipeline some
time ago so that we create the service fresh for each request.

That made me then think, why do we need Clone (and hence buffers) in
AsyncCheckpoint.

This commit is a basically working (two layer tests fail and would need to
be addressed) removal of Clone from AsyncCheckpoint. We could probably
do this to other services which currently require Clone and don't. If we
do that we can probably completely remove the need for "Buffer" from the
plugin codebase.