Entitlement enforcement. (#2596)
Hooked up final entitlement pieces. This won't have any visible affect
on the user as there are no features that currently require entitlement.

Changes:
* Added user facing options. Order of search is ENV, CMD line option,
Registry (API_KEY/GRAPH_REF)
* EntitlementExpander now responsible for expanding entitlements fully,
not just saying if it was warn or update check. This prevents raciness
and helps simplify state machine logic.
* Fixed issue JWT decode issues:
  * UNIX time should be used to decode warn_at and halt_at.
  * Whitespace should be trimmed.
* Hooked final bits in state machine.
* Implemented halt and warning states:
  * Root spans are all error spans.
  * Rate limited log messages indicating that entitlement has expired.
  * Errors reported in metrics.
  * Halt serving requests.



<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: bryn <bryn@apollographql.com>