make argument parsing optional in Executable
if the Executable builder is used, it should not try to parse command
line arguments itself, because the wrapping application may come with
its own set of arguments.
With this change, it is possible to parse command line arguments before
building the Executable, passing the required parameters to the builder.
And the default behaviour is still to parse them directly so this is not
a breaking change.
The main API change here is exposing the `Opt` struct, and adding a
method to the builder