Sample FTV1 by supergraph request instead of by subgraph request (#2656)
Because tracing can be costly, it is only enabled for a configurable
fraction of requests. Each request is selected for tracing or not with a
corresponding probability. This used to be done as part of the subgraph
service, meaning that when a single supergraph request handled by the
Router involves making multiple subgraph requests, it would be possible
(and likely) that tracing would only be enabled for some of those
sub-requests. If this same supergraph request is repeated enough times
the aggregated metrics should be fine, but for smaller sample size this
risks giving an unexpectedly partial view of what’s happening.

Now each supergraph request recieved by the Router is sampled or not for
FTV1, and all corresponding subgraph requests reuse the same result.

Fixes https://github.com/apollographql/router/issues/2655

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`