Simplify ci config (#2684)
This commit reworks the CI config to simplify and speed up the total
runtime.

The use of caching has been reexamined, and now uses the following:

* Daily cache for toolchain and related utilities. Can change
independently of other caches.
* Closest common ancestor to dev for target directory. This means that
when a branch is merged into dev it will take longer to run as it will
be creating caches, but branches off of dev will be quick to run. This
should mean PRs are quick, which is the vast majority of builds.

Removed much platform specific behavior by using bash as the shell for
windows, and doing a little pre-setup to remove differences between
environments.


<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: bryn <bryn@apollographql.com>