Fix visibility of telemetry plugin (#2740)
The telemetry plugin is currently pub. However, since the refactor of
the Telemetry plugin and associated tests this doesn't need to be.

It is not a breaking change to fix this as the plugin was exported under
the _private module which was clearly marked as internal. However, this
can be removed altogether.

Previously we also had the Telemetry struct exported as TelemetryPlugin.
This is no-longer necessary so I renamed the struct for consistency.

---------

Co-authored-by: bryn <bryn@apollographql.com>