Otel shutdown may sometime hang due to `Telemetry::Drop` using a `tokio::blocking_task` to flush the `trace_provider`.
Tokio doesn't finish executing tasks before termination https://github.com/tokio-rs/tokio/issues/1156.
This means that if the runtime is shutdown there is potentially a race where `trace_provider` may not be flushed.
By using a thread it doesn't matter if the tokio runtime is shut down.
This is likely to happen in tests due to the tokio runtime being destroyed when the test method exits.

Fixes #2668