Support aliased top-level `__typename`
PR #1762 implemented support for the query `{ __typename }` but ignored
the possibility that the field might be aliased or repeated. For spec
compatibility (and because this is a convenient schema-agnostic query
for testing purposes), support these cases as well.

Note that the `solve_*` tests of the formatter passed before this
change, as the formatter already processed this correctly.