adding support for specified log "message" / type (#2797)
*Description here*

Fixes #2777. 

The current logs only state the output for "rhai_*", which while mostly
descriptive, doesn't cover the use-case of needing custom logs for
observability. For example:

```rhai
fn subgraph_service(service, subgraph) {
    service.map_response(|response| {
        if response.body.errors != {} {
            log_error(response.body.errors, `Subgraph: ${subgraph}`);
        }
    });
}
```

Will enable customers to have per-subgraph error logging coming from the
router, annotated with the subgraph name without parsing the `out`
value.

Previously, this would only be bucketed under the `rhai_error` message
type.

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`