Fix default Uplink endpoint URLs (#2830)
The Uplink URLs that come with the router were incorrect. The primary
URL worked, but the backup URL did not.

This fixes and adds an integration test to ensure that all Uplink URLs
can be contacted and data retrieved.

Users of older versions of the router should upgrade as soon as
possible.
However, if this is not possible then a workaround is to set
`APOLLO_UPLINK_ENDPOINTS=https://uplink.api.apollographql.com/,https://aws.uplink.api.apollographql.com/`.

Fixes #2827 
<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [x] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

Co-authored-by: bryn <bryn@apollographql.com>