use kubeconform to validate rendered yaml for helm chart  (#2835)
PR: #2824 provided the outline of an approach to validating helm charts
to prevent helm chart errors creeping in.

This PR takes that outline and fits it in with our existing circleci
config. I've also added `check_compliance` back in as a qualifying step
before running tests.

Fixes #1914

Co-authored-by: damienpontifex <damien.pontifex@gmail.com>

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ x ] Changes are compatible[^1]
~- [ ] Documentation[^2] completed~
~- [ ] Performance impact assessed and acceptable~
~- Tests added and passing[^3]~
    ~- [ ] Unit Tests~
    ~- [ ] Integration Tests~
    ~- [ ] Manual Tests~

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`