alter unix_now() to return an i64, not a u64
rhai integers are always i64, so returning a u64 from unix_now() makes
it difficult to work with.

If we truncate the u64 value to be an i64, we lose some representation
of range, but we still have many seconds to work with, so it's the
lesser evil in this situation.