docs: Update the federation version matrix to show support for Federation 2.4.0 (#2889)
This follows up with a docs-only change to the version matrix following
the relese of Router v1.13.1. This PR targets `main`.

Router 1.13.1 now supports using schemas composed with Federation
v2.4.0. However, no new features are yet available in the Router on
account of that Federation version, thus the patch release of the
Router.

Ref: https://github.com/apollographql/router/pull/2872
Ref: https://github.com/apollographql/router/pull/2869
Ref: https://github.com/apollographql/router/pull/2706