Move Apollo telemetry from an opentelemetry exporter to a tracing layer (#2908)
Apollo telemetry was running as an Opentelemetry exporter, but this came with performance and reliability drawbacks. In particular, the exporters work by accumulating all spans of a trace, and all their attributes, then process them all at once when the request ends. Also, the late end for some span due to field tracing, required that we store spans in a central LRU cache.

This moves Apollo telemetry to a tracing-subscriber layer, where it can pick only the data it needs, and store it in a per trace map, cleaning it up quickly on request end. Also, this leverages the batching task used by the metrics exporter