Update to apollo-parser@0.5.1 and apollo-compiler@0.8.0 (#2958)
The parser now does less recursion so our recursion limit test needs a
change. Fixes CI failure with updated dependencies like
https://app.circleci.com/pipelines/github/apollographql/router/11672/workflows/a57c04c5-6849-40d3-b40a-437a543e586d/jobs/76934?invite=true#step-117-1189

Cherry-picked from a larger branch.

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

Co-authored-by: Irina Shestak <shestak.irina@gmail.com>