Re-structure the rhai logging output
fixes: #2777

The Rhai logging output was "standardized" about a year ago. Since that
time, significant efforts have been made with the rest of the router to
standardize on a slightly different format. This PR brings rhai log
output into line with the rest of the router.

It also addresses the requirement to make the "message" output
specifiable by the script author.