respect studio's backoff feedback
Studio may return a 429 with a `RETRY_AFTER` header containing a number
of seconds before trying to re-submit traces.

This implementation makes a few assumptions:
 - firstly, I'm assuming the RETRY_AFTER duration is seconds
 - secondly, rather than buffering for later submission, I'm throwing
   away the affected trace reports.

The first assumption can be easily assessed for accuracy. The second one
is more more opinionated. I'm working on the basis that buffering unsent
traces in the router is a "bad thing". We could bound the number of
unsent traces and try to resubmit them after the RETRY_AFTER period
expires, but that's more work and could always be added later as a
refinement to this simple approach.

Summary: This is a draft for discussion of the second assumption.