Filter spans before sending them to the opentelemetry layer (#2894)
The Opentelemetry layer accumulates data from all spans before sending
it to the opentelemetry span exporters. This means that sampling
decisions are done after a lot of data was already allocated. And that
this work is done even if no exporter is set up.

This adds a sampling filter around the opentelemetry layer, to do the
sampling decision as early as possible, and avoid this work entirely if
no exporter is configured