Ensure `--prerelease` flag is passed to `gh create release` for pre-releases
This changes our script to be explicit and pass the `--prerelease` version
to our increasingly automated release pipeline when the `-` character is
detected in the "version" property of the release being published.

Typically, GitHub marks "latest" versions "based on semver"[1], but GitHub
still marks "pre-release versions" (as defined by Semver 2.0 [2]) as
"latest" irregardless of that meaning to semver.

This seems a bit unexpected, but it is the way it is, so we'll account for
it by making sure we're explicit.

[1]: https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository#:~:text=the%20latest%20release%20label%20will%20automatically%20be%20assigned%20based%20on%20semantic%20versioning
[2]: https://semver.org/#spec-item-9

Follows-up: #3003 (root-cause) (root-cause) (root-cause) (root-cause)